pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hermesd"
version = "0.1.0"
description = "TUI monitoring dashboard for Hermes AI agent"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Nous Research"},
    {name = "Ion Mudreac", email = "mudreac@gmail.com"},
]
keywords = ["hermes", "agent", "dashboard", "tui", "monitoring", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "rich>=14.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/NousResearch/hermesd"
Repository = "https://github.com/NousResearch/hermesd"
Issues = "https://github.com/NousResearch/hermesd/issues"
Changelog = "https://github.com/NousResearch/hermesd/blob/main/CHANGELOG.md"

[project.scripts]
hermesd = "hermesd.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.sdist]
include = ["hermesd/", "tests/", "LICENSE", "README.md", "CHANGELOG.md"]
